Release

+++ packages/desktop-ui-lib/src/lib/settings/settings.component.ts
@@ -5,17 +5,21 @@ import {
 	ViewChild,
 	ElementRef,
 	NgZone,
-	AfterViewInit
+	AfterViewInit,
+	Optional
 } from '@angular/core';
 import { TimeTrackerService } from '../time-tracker/time-tracker.service';
 import { NbDialogService, NbToastrService } from '@nebular/theme';
 import { ElectronService } from '../electron/services';
-import { BehaviorSubject, Observable, filter, tap } from 'rxjs';
+import { BehaviorSubject, Observable, filter, tap, firstValueFrom } from 'rxjs';
 import { AboutComponent } from '../dialogs/about/about.component';
 import { SetupService } from '../setup/setup.service';
 import * as moment from 'moment';
 import { ToastrNotificationService } from '../services';
+import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
+import { AuthStrategy } from '../auth';
 
+@UntilDestroy({ checkProperties: true })
 @Component({
 	selector: 'ngx-settings',
 	templateUrl: './settings.component.html',
@@ -49,6 +53,7 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		this._logContents$.next(logs);
 	}
 	logIsOpen = false;
+	private _isRestart$: BehaviorSubject<boolean>;
 
 	appName: string = this.electronService.remote.app.getName();
 	menus = [];
@@ -379,7 +384,7 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 	autoLaunch = null;
 	minimizeOnStartup = null;
 	authSetting = null;
-	currentUser$: BehaviorSubject<any> = new BehaviorSubject({});
+	currentUser$: BehaviorSubject<any> = new BehaviorSubject(null);
 	serverTypes = {
 		integrated: 'Integrated',
 		custom: 'Custom',
@@ -426,7 +431,9 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		private toastrService: NbToastrService,
 		private _dialogService: NbDialogService,
 		private _setupService: SetupService,
-		private _notifier: ToastrNotificationService
+		private _notifier: ToastrNotificationService,
+		@Optional()
+		private _authStrategy: AuthStrategy
 	) {
 		this._loading$ = new BehaviorSubject(false);
 		this._automaticUpdate$ = new BehaviorSubject(false);
@@ -454,6 +461,7 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		this._restartDisable$ = new BehaviorSubject(false);
 		this._isHidden$ = new BehaviorSubject(true);
 		this._simpleScreenshotNotification$ = new BehaviorSubject(false);
+		this._isRestart$ = new BehaviorSubject(false);
 	}
 
 	ngOnInit(): void {
@@ -463,14 +471,24 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 			.pipe(
 				tap(({ status }) => this._restartDisable$.next(!status)),
 				filter(() => !this._isCheckHost.status),
-				tap(() => this._restartDisable$.next(true))
+				tap(() => this._restartDisable$.next(true)),
+				untilDestroyed(this)
 			)
 			.subscribe();
 		this.isCheckHost$
 			.pipe(
 				tap(({ status }) => this._restartDisable$.next(!status)),
 				filter(() => !this._isConnectedDatabase.status),
-				tap(() => this._restartDisable$.next(true))
+				tap(() => this._restartDisable$.next(true)),
+				untilDestroyed(this)
+			)
+			.subscribe();
+		this.isRestart$
+			.pipe(
+				tap((isRestart: boolean) =>
+					this._restartDisable$.next(isRestart)
+				),
+				untilDestroyed(this)
 			)
 			.subscribe();
 	}
@@ -515,18 +533,18 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 				});
 				this._simpleScreenshotNotification$.next(setting?.simpleScreenshotNotification);
 				this.selectedPeriod = setting?.timer?.updatePeriod;
-				if (!this.isServer) {
+				if (this.isDesktopTimer) {
 					await this.getUserDetails();
 				}
 				this.menus = this.isServer
 					? ['Update', 'Advanced Setting', 'About']
 					: [
-							...(auth && auth.allowScreenshotCapture ? ['Screen Capture'] : []),
-							'Timer',
-							'Update',
-							'Advanced Setting',
-							'About'
-					  ];
+						...(auth && auth.allowScreenshotCapture ? ['Screen Capture'] : []),
+						'Timer',
+						'Update',
+						'Advanced Setting',
+						'About'
+					];
 				const lastMenu =
 					this._selectedMenu && this.menus.includes(this._selectedMenu) ? this._selectedMenu : this.menus[0];
 				this._selectedMenu$.next(lastMenu);
@@ -669,8 +687,8 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		});
 
 		this.electronService.ipcRenderer.on('_logout_quit_install_', (event, arg) => {
-			this._ngZone.run(() => {
-				this.logout(true);
+			this._ngZone.run(async () => {
+				await this.restartAndUpdate();
 			});
 		});
 	}
@@ -786,11 +804,12 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		this.updateSetting(value, 'prerelease');
 	}
 
-	restartApp() {
-		if (this.isServer && this.serverIsRunning) {
-			this._restartDisable$.next(true);
-		} else {
-			this.logout();
+	public async restartApp(): Promise<void> {
+		this._isRestart$.next(true);
+		if (!this.isServer && !this.authSetting.isLogout) {
+			await firstValueFrom(this._authStrategy.logout());
+			this.currentUser$.next(null);
+			localStorage.clear();
 		}
 		const thConfig = {};
 		this.thirdPartyConfig.forEach((item) => {
@@ -850,8 +869,8 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 		this.electronService.ipcRenderer.send('check_for_update');
 	}
 
-	restartAndUpdate() {
-		this.logout(true);
+	public async restartAndUpdate(): Promise<void> {
+		await this.logout(true);
 	}
 
 	toggleAwView(value) {
@@ -895,10 +914,26 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 	/*
 	 * Logout desktop timer
 	 */
-	logout(isAfterUpgrade?: boolean) {
-		console.log('On Logout');
-		localStorage.clear();
-		this.electronService.ipcRenderer.send('logout_desktop', isAfterUpgrade);
+	public async logout(isAfterUpgrade?: boolean): Promise<void> {
+		let isLogout = true;
+
+		if (this.appSetting?.timerStarted) {
+			isLogout = await this.electronService.ipcRenderer.invoke(
+				'LOGOUT_STOP'
+			);
+		} else {
+			localStorage.clear();
+		}
+
+		if (isLogout) {
+			console.log('On Logout');
+			this.electronService.ipcRenderer.send(
+				'logout_desktop',
+				isAfterUpgrade
+			);
+		} else {
+			console.log('Logout Cancelled');
+		}
 	}
 
 	onServerChange(val) {
@@ -954,6 +989,7 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 				break;
 		}
 		this.toastrService.show(message, `Success`, { status: arg.status });
+		this._isRestart$.next(false);
 	}
 
 	logBoxChange(e) {
@@ -1171,5 +1207,9 @@ export class SettingsComponent implements OnInit, AfterViewInit {
 
 	public get simpleScreenshotNotification(): boolean {
 		return this._simpleScreenshotNotification$.getValue();
+	}
+
+	public get isRestart$(): Observable<boolean> {
+		return this._isRestart$.asObservable();
 	}
 }
